+++ src/main/web/report/awards/edit-awards-presenters.js
@@ -1,0 +1,6 @@
+"use strict";
+
+
+document.addEventListener('DOMContentLoaded', () => {
+    pageInit();
+});